backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.32.1
anthropic>=0.40.0
google-genai>=1.0.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
scalar-fastapi>=1.0.3
